# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hush omaded",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "c47411f6de96b5b30018c647e449e470",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "patronymic",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Shahboz"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Qurbonov"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "G`aybulloevich"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "http://example.com/shahboz.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован"
                    },
                    "400": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "82d74137970376a374510566964c9813",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "000001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получение информации о пользователе",
                "operationId": "3e20ff565e17a76fa1384c8a31dee558",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Shahboz"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "Qurbonov"
                                        },
                                        "patronymic": {
                                            "type": "string",
                                            "example": "G`aybulloevich"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "000001"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://example.com/avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновление профиля пользователя",
                "operationId": "33262d5a04c347508a6ca97b80e99451",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "avatar"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Shahboz"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Qurbonov"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "G`aybulloevich"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "http://example.com/avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профиль успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Shahboz"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "Qurbonov"
                                        },
                                        "patronymic": {
                                            "type": "string",
                                            "example": "G`aybulloevich"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "000001"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://example.com/avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{search}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Поиск пользователей",
                "description": "Поиск пользователей по ID, имени, фамилии, отчеству, логину или дате создания",
                "operationId": "getUserName",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Введите ID, имя, фамилию, отчество, логин или дату создания пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2025-02-12"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Найденные пользователи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Шахбоз"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Курбанов"
                                            },
                                            "patronymic": {
                                                "type": "string",
                                                "example": "Гайбуллоевич"
                                            },
                                            "login": {
                                                "type": "string",
                                                "example": "shahboz123"
                                            },
                                            "avatar": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "http://example.com/avatar.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-02-12 10:23:53"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Создание чата",
                "operationId": "fcdf2c6b221e1f5cb6eebba8feda11fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "support",
                                            "private"
                                        ],
                                        "example": "private"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Чат успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "enum": [
                                                "support",
                                                "private"
                                            ],
                                            "example": "private"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{id}": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Получение данных о чате",
                "operationId": "cef4a1bfa10132b70f43ac02c00e1b8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщения чата",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Привет!"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "example": "http://example.com/file.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-02-06T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Чат не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}